+++ src/jobs/healthCheck.ts
@@ -1,10 +1,10 @@
 import { getConfig } from '../config/config';
 import { Boxes } from '../../src/ergo/boxes';
 import { HealthCheckSingleton } from '../../src/utils/healthCheck';
-import WinstonLogger from '@rosen-bridge/winston-logger';
+import { DefaultLoggerFactory } from '@rosen-bridge/abstract-logger';
 import { ERGO_CHAIN_NAME } from '../config/constants';
 
-const logger = WinstonLogger.getInstance().getLogger(import.meta.url);
+const logger = DefaultLoggerFactory.getInstance().getLogger(import.meta.url);
 
 /**
  * updates the permit thresholds based on needed Rwts for each commitment