+++ src/jobs/initScanner.ts
@@ -2,7 +2,7 @@ import { CardanoOgmiosScanner, GeneralScanner } from '@rosen-bridge/scanner';
 import * as Constants from '../config/constants';
 import { getConfig } from '../config/config';
 import { scanner } from '../utils/scanner';
-import WinstonLogger from '@rosen-bridge/winston-logger';
+import { DefaultLoggerFactory } from '@rosen-bridge/abstract-logger';
 import { EvmRpcScanner } from '@rosen-bridge/evm-rpc-scanner';
 
 const allConfig = getConfig();
@@ -14,7 +14,7 @@ const {
   binance: binanceConfig,
 } = allConfig;
 
-const logger = WinstonLogger.getInstance().getLogger(import.meta.url);
+const logger = DefaultLoggerFactory.getInstance().getLogger(import.meta.url);
 
 const scanningJob = async (interval: number, scanner: GeneralScanner<any>) => {
   try {