+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # @rosen-bridge/watcher
 
+## 4.0.2
+
+### Patch Changes
+
+- Fix redeem job to redeem the invalid commitment after the reward transaction
+
+## 4.0.1
+
+### Patch Changes
+
+- Add eRSN token info to assets and revenue apis
+- Update logger structure
+
 ## 4.0.0
 
 ### Major Changes